Visibility enhancements for partners (#3744)
Closes #3737 and Closes #3733. This adds a new visibility setting of
`applicant partners`. This addition turned out to be a bit unwieldy
because of how situation it is. The main constraints I was designing for
was in the case of a lab/fund that was using the partner role for
vendors. I was trying to also account for the possibility that multiple
vendors will be assigned to a single submission (which I believe is an
edge case).

The main concern was applicants/vendors inputting sensitive data into
the communications tab and having each other see such. For example: if a
vendor uploads banking information an applicant (or other vendor) should
not be able to see this information and vice versa.

Also tried to include a few docstring additions in lieu of the
documentation updates.