Swapped out hamburger menu for a `Login` button in mobile view (#3812)
Fixes #3811. This removes the hamburger menu for unauthenticated users
in favor of a simple `Login` button. If a signup is intended and
`ENABLE_PUBLIC_SIGNUP` is enabled on the instance, the signup button can
be found on the `/auth/` view. I didn't want to overload the mobile view
by putting login & signup next to eachother.